Add ChairController
Changing the player pawn to the chair does not work because Money expects the local pawn to be Player. Moving the money to an EntityComponent of the Client would maybe be better, but doing so would break other people's stuff. There's some camera and use code that runs on the Player pawn that interferes with what I'm trying to do, so I put in bools to toggle such behavior.

It's hacky. I'll think of a better way before merging back in to main.